+++ packages/feature-flags/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "flag-editor",
+  "name": "feature-flags",
   "version": "1.0.0",
   "main": "index.js",
   "license": "MIT",
